#273 Added config entries.

+++ petblocks-api/src/main/kotlin/com/github/shynixn/petblocks/api/persistence/entity/AIParticle.kt
@@ -14,5 +14,5 @@ interface AIParticle : AIBase {
     /**
      * Amount of seconds until the next particle is played.
      */
-    var delayBetweenPlaying: Int
+    var delayBetweenPlaying: Double
 }
